@@ -60,29 +60,44 @@ namespace starskytest.Helpers
 			var swaggerFilePath = Path.Join(_appSettings.TempFolder, _appSettings.Name + ".json");
 
 			FilesHelper.DeleteFile(swaggerFilePath);
-			
+			System.Console.WriteLine("swaggerFilePath " + swaggerFilePath);
+
 			var host = WebHost.CreateDefaultBuilder()
 				.UseUrls("http://localhost:5051")
 				.ConfigureServices(services =>
 				{
-					services.AddMvcCore().AddApiExplorer(); // use core and AddApiExplorer to make it faster
-					// https://offering.solutions/blog/articles/2017/02/07/difference-between-addmvc-addmvcore/
+
+#if NETCOREAPP3_0
+					services.AddMvcCore().AddApiExplorer();
+						//.AddNewtonsoftJson();
+#else
+	services.AddMvcCore().AddApiExplorer(); // use core and AddApiExplorer to make it faster
+		// https://offering.solutions/blog/articles/2017/02/07/difference-between-addmvc-addmvcore/
+#endif
+
 					services.AddSwaggerGen();
 					new SwaggerHelper(_appSettings).Add01SwaggerGenHelper(services);
 
 				})
 				.Configure(app =>
 				{
-//					app.UseStaticFiles(); // disabled for now
-					app.UseMvc();
-					app.UseMvc(routes =>
+
+#if NETCOREAPP3_0
+					app.UseRouting();
+					app.UseEndpoints(endpoints =>
 					{
-						routes.MapRoute(
-							name: "default",
-							template: "{controller=Home}/{action=Index}/{id?}");
+						endpoints.MapControllerRoute("default", "{controller=Home}/{action=Index}/{id?}");
 					});
-					new SwaggerHelper(_appSettings).Add02AppUseSwaggerAndUi(app); 
-					new SwaggerHelper(_appSettings).Add03AppExport(app); 
+#else
+	        app.UseMvc(routes =>
+					 {
+						 routes.MapRoute(
+							 name: "default",
+							 template: "{controller=Home}/{action=Index}/{id?}");
+					 });
+#endif
+					new SwaggerHelper(_appSettings).Add02AppUseSwaggerAndUi(app);
+					new SwaggerHelper(_appSettings).Add03AppExport(app);
 				}).Build();
 
 			await host.StartAsync();
@@ -91,7 +106,9 @@ namespace starskytest.Helpers
 			Assert.AreEqual(true,FilesHelper.ExistFile(swaggerFilePath));
 
 			var swaggerFileContent = new PlainTextFileHelper().ReadFile(swaggerFilePath);
-				
+
+			System.Console.WriteLine("swaggerFileContent " + swaggerFileContent);
+
 			Assert.AreEqual(true, swaggerFileContent.Contains($"\"title\": \"{_appSettings.Name}\""));
 			
 			
