@@ -5,9 +5,8 @@ using System.Globalization;
 using System.IO;
 using System.Linq;
 using System.Text;
+using System.Text.Json.Serialization;
 using System.Text.RegularExpressions;
-using Newtonsoft.Json;
-using Newtonsoft.Json.Converters;
 using starskycore.Helpers;
 
 namespace starskycore.Models
@@ -69,8 +68,9 @@ namespace starskycore.Models
         public DateTime DateTime{ get; set; }
 	    
 	    [NotMapped]
-	    [JsonConverter(typeof(StringEnumConverter))]
-	    public ImportStatus Status { get; set; }
+	    //[JsonConverter(typeof(StringEnumConverter))]
+		[JsonConverter(typeof(JsonStringEnumConverter))]
+		public ImportStatus Status { get; set; }
 	    
 	    [NotMapped]
 		public FileIndexItem FileIndexItem { get; set; }
@@ -316,8 +316,7 @@ namespace starskycore.Models
 
                     if (createFolder)
                     {
-                        Console.WriteLine("childFullDirectory");
-                        Console.WriteLine(childFullDirectory);
+						Console.Write("+");
                         Directory.CreateDirectory(childFullDirectory);
                     }
                 }
